+++ src/test/kotlin/com/github/zero9178/mlirods/SemanticTokensAnnotatorTest.kt
@@ -17,7 +17,6 @@ class SemanticTokensAnnotatorTest : BasePlatformTestCase() {
             }
             def Foo : Foo {
                 let <text_attr textAttributesKey="TABLEGEN_KEYWORD">append</text_attr> <text_attr textAttributesKey="TABLEGEN_FIELD">s</text_attr> = "";
-                let <error descr="Expected one of 'append' or 'prepend' instead of 'apend'" textAttributesKey="ERRORS_ATTRIBUTES">apend</error> <text_attr textAttributesKey="TABLEGEN_FIELD">s</text_attr> = "";
             }
             defvar test = Foo.<text_attr textAttributesKey="TABLEGEN_FIELD">s</text_attr>;
             #define <text_attr textAttributesKey="TABLEGEN_PREPROCESSOR_MACRO_NAME">FOO</text_attr>